refactor(Hamburger): add Hamburger folder
