+++ frontend/src/components/Header/Hamburger/Hamburger.module.css
@@ -1,0 +1,56 @@
+.hamburger {
+  --width: 1.7rem;
+  --height: var(--width);
+
+  display: flex;
+  display: none;
+  flex-direction: column;
+  justify-content: space-evenly;
+  width: var(--width);
+  height: var(--height);
+
+  background: none;
+  border: none;
+  cursor: pointer;
+}
+
+.line {
+  background-color: var(--light-blue);
+  border-radius: 0.1rem;
+  height: 0.2rem;
+  width: 100%;
+
+  transition: var(--ease-out-quint);
+}
+
+.line:nth-child(1) {
+  transform-origin: right top;
+}
+
+.lineCross:nth-child(1) {
+  rotate: -45deg;
+  translate: -0.4rem -0.1rem;
+}
+
+.line:nth-child(2) {
+  opacity: 1;
+}
+
+.lineCross:nth-child(2) {
+  opacity: 0;
+}
+
+.line:nth-child(3) {
+  transform-origin: right bottom;
+}
+
+.lineCross:nth-child(3) {
+  rotate: 45deg;
+  translate: -0.4rem 0.1rem;
+}
+
+@media (max-width: 630px) {
+  .hamburger {
+    display: flex;
+  }
+}
