+++ frontend/src/components/Header/Hamburger/Hamburger.tsx
@@ -1,14 +1,14 @@
 import { useContext } from "react";
 
-import { NavContext } from "../../context/NavContext";
+import { NavContext } from "../../../context/NavContext";
 
-import styles from "./Header.module.css";
+import styles from "./Hamburger.module.css";
 
 const Hamburger = () => {
   const nav = useContext(NavContext);
 
-  const lineClassNames = [styles.hamburgerLine];
-  if (nav.isOpened) lineClassNames.push(styles.hamburgerLineActive);
+  const lineClassNames = [styles.line];
+  if (nav.isOpened) lineClassNames.push(styles.lineCross);
 
   return (
     <button className={styles.hamburger} onClick={nav.toggle}>
