+++ frontend/src/components/Header/Header.module.css
@@ -50,57 +50,6 @@
   gap: 0.5rem;
 }
 
-.hamburger {
-  --width: 1.7rem;
-  --height: var(--width);
-
-  display: flex;
-  display: none;
-  flex-direction: column;
-  justify-content: space-evenly;
-  width: var(--width);
-  height: var(--height);
-
-  background: none;
-  border: none;
-  cursor: pointer;
-}
-
-.hamburgerLine {
-  background-color: var(--light-blue);
-  border-radius: 0.1rem;
-  height: 0.2rem;
-  width: 100%;
-
-  transition: var(--ease-out-quint);
-}
-
-.hamburgerLine:nth-child(1) {
-  transform-origin: right top;
-}
-
-.hamburgerLineActive:nth-child(1) {
-  rotate: -45deg;
-  translate: -0.4rem -0.1rem;
-}
-
-.hamburgerLine:nth-child(2) {
-  opacity: 1;
-}
-
-.hamburgerLineActive:nth-child(2) {
-  opacity: 0;
-}
-
-.hamburgerLine:nth-child(3) {
-  transform-origin: right bottom;
-}
-
-.hamburgerLineActive:nth-child(3) {
-  rotate: 45deg;
-  translate: -0.4rem 0.1rem;
-}
-
 @media (max-width: 800px) {
   .header {
     gap: 1.5rem;
@@ -177,10 +126,6 @@
     font-size: 1.5rem;
     cursor: pointer;
     font-family: inherit;
-  }
-
-  .hamburger {
-    display: flex;
   }
 
   .dimming {
